pyproject.toml
[project]
name = "enginimate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "docutils>=0.22.2",
    "jinja2>=3.1.6",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.1",
    "langsmith>=0.4.38",
    "manim>=0.19.0",
    "pydantic>=2.12.3",
    "pypandoc>=1.15",
    "python-dotenv>=1.2.1",
    "unstructured[rst]>=0.18.15",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "nest-asyncio>=1.6.0",
    "ruff>=0.14.3",
]

[tool.setuptools.packages]
# Tells setuptools to look for packages starting at the current directory (".")
# and explicitly include the list of folders you care about.
find = { include = ["backend*"] }

[tool.ruff]
lint.select = ["I"]
